requirements.txt
amqp==5.1.1
aniso8601==7.0.0
arabic-reshaper==3.0.0
asgiref==3.7.2
asn1crypto==1.5.1
async-timeout==4.0.2
beautifulsoup4==4.9.1
billiard==4.1.0
braintree==4.6.0
CacheControl==0.13.1
celery==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.3
contextlib2==21.6.0
crispy-bootstrap4==2022.1
cron-descriptor==1.4.0
cryptography==41.0.3
cssselect2==0.7.0
dateparser==1.1.8
defusedxml==0.7.1
diff-match-patch==20230430
Django==4.2.3
django-allauth==0.54.0
django-autoslug==1.9.9
django-bootstrap4==23.2
django-braces==1.15.0
django-celery-beat==2.5.0
django-celery-results==2.5.1
django-ckeditor==6.6.1
django-cors-headers==4.1.0
django-countries==7.5.1
django-crispy-forms==2.0
django-debug-toolbar==4.1.0
django-discover-runner==1.0
django-environ==0.10.0
django-extensions==3.2.3
django-file-form==3.5.2
django-filter==23.2
django-import-export==3.2.0
django-js-asset==2.1.0
django-model-utils==4.3.1
django-mptt==0.14.0
django-mysql==4.10.0
django-redis-cache==3.0.1
django-rename-app==0.1.6
django-role-permissions==3.2.0
django-social-share==2.3.0
django-tables2==2.6.0
django-taggit==4.0.0
django-timezone-field==5.1
django-tinymce==3.6.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
et-xmlfile==1.1.0
fake-factory==9999.9.9
Faker==18.13.0
future==0.18.3
graphene==2.1.9
graphql-core==2.3.2
graphql-relay==2.0.1
greenlet==2.0.2
html5lib==1.1
idna==3.4
ipaddr==2.2.0
jdcal==1.4.1
jet-bridge-base==1.7.19
jet-django==1.7.19
kombu==5.3.1
lockfile==0.12.2
lxml==4.9.3
Markdown==3.4.3
MarkupPy==1.14
msgpack==1.0.5
#mysqlclient==2.2.0
oauthlib==3.2.2
odfpy==1.4.1
openpyxl==3.1.2
oscrypto==1.3.0
packaging==23.1
pep517==0.13.0
Pillow==10.0.0
progress==1.6
promise==2.3
prompt-toolkit==2.0.9
pycodestyle==2.10.0
pycparser==2.21
pyHanko==0.19.0
pyhanko-certvalidator==0.23.0
PyJWT==2.7.0
pyparsing==3.1.0
pypdf==3.12.0
PyPDF2==3.0.1
pypng==0.20220715.0
python-bidi==0.4.2
python-crontab==2.7.1
python-dateutil==2.8.2
python-slugify==8.0.1
python3-openid==3.2.0
pytoml==0.1.21
pytz==2023.3
PyYAML==6.0
qrcode==7.4.2
redis==3.5.3
regex==2023.6.3
reportlab==3.6.13
requests==2.31.0
requests-oauthlib==1.3.1
retrying==1.3.4
Rx==1.6.3
sentry-sdk==1.27.1
six==1.16.0
social-auth-core==4.4.2
soupsieve==2.4.1
SQLAlchemy==1.4.49
sqlparse==0.4.4
sslcommerz-lib==1.0
svglib==1.5.1
tablib==3.5.0
text-unidecode==1.3
tinycss2==1.2.1
toml==0.10.2
tomli==2.0.1
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uritools==4.0.1
urllib3==2.0.3
vine==5.0.0
wcwidth==0.2.6
webencodings==0.5.1
whitenoise==6.5.0
xhtml2pdf==0.2.11
xlrd==2.0.1
xlwt==1.3.0
